完成ekf接入，对齐流程直到重投影和状态建模预测

+++ applications/auto_aim/AutoAimProvider/yolos/yolo11.cpp
@@ -258,7 +258,7 @@ void YOLO11::draw_detections(
 void YOLO11::save(const Armor & armor) const
 {
   auto file_name = fmt::format("{:%Y-%m-%d_%H-%M-%S}", std::chrono::system_clock::now());
-  auto img_path = fmt::format("{}/{}_{}.jpg", save_path_, armor.name, file_name);
+  auto img_path = fmt::format("{}/{}_{}.jpg", save_path_, static_cast<int>(armor.name), file_name);
   cv::imwrite(img_path, tmp_img_);
 }
